Crate rustache

The main crate for the Rustache library.

Rustache is a flexible template engine for Rust.

Structs

HashBuilder
HashBuilder is a helper type that constructs Data types in a HashMap

VecBuilder
VecBuilder is a helper type that constructs Data types in a Vector

Enums

RustacheError
Enum to handle errors from the Rustache library.

Traits

Render
Defines a renderable trait, so that all of our data is renderable

Functions

render_file
Render a template from the given template file

render_text
Render the given template string

Type Definitions

RustacheResult
Alias for Result