Struct rustache::HashBuilder

pub struct HashBuilder<'a> {
    // some fields omitted
}

HashBuilder is a helper type that constructs Data types in a HashMap

Methods

impl<'a> HashBuilder<'a>

fn new() -> HashBuilder<'a>

Create a new HashBuilder instance

fn insert_string<K: ToString, V: ToString>(self, key: K, value: V) -> HashBuilder<'a>

Add a String to the HashBuilder

use rustache::HashBuilder;
let data = HashBuilder::new()
    .insert_string("game", "Hearthstone: Heroes of Warcraft");

fn insert_bool<K: ToString>(self, key: K, value: bool) -> HashBuilder<'a>

Add a Boolean to the HashBuilder

use rustache::HashBuilder;
let data = HashBuilder::new()
    .insert_bool("playing", true);

fn insert_int<K: ToString>(self, key: K, value: i32) -> HashBuilder<'a>

Add an Integer to the HashBuilder

use rustache::HashBuilder;
let data = HashBuilder::new()
    .insert_int("age", 10i32)
    .insert_int("drinking age", -21i32);

fn insert_float<K: ToString>(self, key: K, value: f64) -> HashBuilder<'a>

Add a Float to the HashBuilder

use rustache::HashBuilder;
let data = HashBuilder::new()
    .insert_float("pi", 3.141596f64)
    .insert_float("phi", 1.61803398875f64);

fn insert_vector<F: Fn(VecBuilder<'a>) -> VecBuilder<'a>, K: ToString>(self, key: K, f: F) -> HashBuilder<'a>

Add a Vector to the HashBuilder

use rustache::HashBuilder;
let data = HashBuilder::new()
    .insert_vector("classes", |builder| {
        builder
            .push_string("Mage".to_string())
            .push_string("Druid".to_string())
    });

fn insert_hash<F: Fn(HashBuilder<'a>) -> HashBuilder<'a>, K: ToString>(self, key: K, f: F) -> HashBuilder<'a>

Add a Hash to the HashBuilder

use rustache::HashBuilder;
let data = HashBuilder::new()
    .insert_hash("hero1", |builder| {
        builder
            .insert_string("first_name", "Anduin")
            .insert_string("last_name", "Wrynn")
    })
    .insert_hash("hero2", |builder| {
        builder
            .insert_string("first_name", "Jaina")
            .insert_string("last_name", "Proudmoore")
    });

fn insert_lambda<K: ToString>(self, key: K, f: &'a mut FnMut(String) -> String) -> HashBuilder<'a>

Add a Lambda that accepts a String and returns a String to the HashBuilder

use rustache::HashBuilder;
let mut f = |_| { "world".to_string() };
let data = HashBuilder::new()
    .insert_lambda("lambda", &mut f);

fn set_partials_path(self, path: &'a str) -> HashBuilder<'a>

Set a path to partials data