Crate rustache

The main crate for the Rustache library.

Rustache is a flexible template engine for Rust.

Structs

HashBuilder
HashBuilder is a helper type that constructs Data types in a HashMap

VecBuilder
VecBuilder is a helper type that constructs Data types in a Vector

Functions

render_file_from_hb
Render a template file from a HashBuilder to a specified writer

render_file_from_json_enum
Render a template file from a Rust JSON Enum to a specified writer

render_file_from_json_file
Render a template file from a JSON file to a specified writer

render_file_from_json_string
Render a template file from a JSON string to a specified writer

render_text_from_hb
Render template text from a HashBuilder to a specified writer

render_text_from_json_enum
Render template text from a Rust JSON Enum to a specified writer

render_text_from_json_file
Render template text from a JSON file to a specified writer

render_text_from_json_string
Render template text from a JSON string to a specified writer